Updates to styles & a few bug fixes
