+++ src/ts/index.tsx
@@ -4,7 +4,7 @@
 // import { AuthProvider } from "./ui/AuthProvider";
 // import { theme } from "./ui/theme";
 // import { render } from "solid-js/web";
-import "./initJSX";
+import "./ui/compiler/jsx";
 import { Application } from "./ui/application/Application";
 import { RouterProvider } from "./ui/common/RouterProvider";
