+++ src/ts/ui/common/Avatar.tsx
@@ -16,7 +16,7 @@ export class Avatar extends Component<Props> {
     };
   }
 
-  css() {
+  getStyle() {
     return css`
       :host {
         position: relative;
