+++ src/ts/ui/common/Date.tsx
@@ -9,17 +9,15 @@ interface Props {
 @register("x-date")
 export class Date extends Component<Props> {
   init() {
-    return () => {
-      return (
-        <div>
-          {this.props.date
-            ? format(
-                this.props.date,
-                this.props.withTime || this.props.withTime === undefined ? "do MMMM y, p" : "do MMMM y"
-              )
-            : "-"}
-        </div>
-      );
-    };
+    return () => (
+      <div>
+        {this.props.date
+          ? format(
+              this.props.date,
+              this.props.withTime || this.props.withTime === undefined ? "do MMMM y, p" : "do MMMM y"
+            )
+          : "-"}
+      </div>
+    );
   }
 }
