+++ src/ts/ui/common/Modal.tsx
@@ -20,45 +20,43 @@ export class Modal extends Component<Props> {
   }
 
   init() {
-    return () => {
-      const handleCancel = (e: MouseEvent) => {
-        this.props.onCancel && this.props.onCancel();
-        this.props.onClose && this.props.onClose();
-      };
-
-      const handleOk = (e: MouseEvent) => {
-        this.props.onOk && this.props.onOk();
-        this.props.onClose && this.props.onClose();
-      };
-
-      return (
-        <Popup open={this.props.open} onClose={this.props.onClose} withBackground={this.props.withBackground}>
-          {typeof this.props.title === "string" ? (
-            <Typography variant="h2">{this.props.title}</Typography>
-          ) : (
-            this.props.title || ""
-          )}
-          <div class="content">
-            <slot></slot>
-          </div>
-          {this.props.hideConfirmButtons ? (
-            ""
-          ) : (
-            <div class="button-container">
-              <Button onClick={handleCancel} class="cancel" variant="outlined">
-                Cancel
-              </Button>
-              <Button onClick={handleOk} class="ok">
-                Okay
-              </Button>
-            </div>
-          )}
-        </Popup>
-      );
+    const handleCancel = (e: MouseEvent) => {
+      this.props.onCancel && this.props.onCancel();
+      this.props.onClose && this.props.onClose();
     };
+
+    const handleOk = (e: MouseEvent) => {
+      this.props.onOk && this.props.onOk();
+      this.props.onClose && this.props.onClose();
+    };
+
+    return () => (
+      <Popup open={this.props.open} onClose={this.props.onClose} withBackground={this.props.withBackground}>
+        {typeof this.props.title === "string" ? (
+          <Typography variant="h2">{this.props.title}</Typography>
+        ) : (
+          this.props.title || ""
+        )}
+        <div class="content">
+          <slot></slot>
+        </div>
+        {this.props.hideConfirmButtons ? (
+          ""
+        ) : (
+          <div class="button-container">
+            <Button onClick={handleCancel} class="cancel" variant="outlined">
+              Cancel
+            </Button>
+            <Button onClick={handleOk} class="ok">
+              Okay
+            </Button>
+          </div>
+        )}
+      </Popup>
+    );
   }
 
-  css() {
+  getStyle() {
     return css`
       .button-container {
         text-align: right;
